+++ BLL/DTOModels/ResponseDTO/OrderPositionResponseDTO.cs
@@ -9,11 +9,11 @@ namespace BLL.DTOModels.ResponseDTO
     public class OrderPositionResponseDTO
     {
         public string ProductName { get; }
-        public double Price { get; }
+        public decimal Price { get; }
         public int Amount { get; }
-        public double Value { get; }
+        public decimal Value { get; }
 
-        public OrderPositionResponseDTO(string productName, double price, int amount, double value)
+        public OrderPositionResponseDTO(string productName, decimal price, int amount, decimal value)
         {
             ProductName = productName;
             Price = price;
